chore: pg execute sql one at a time (#12443)
Former-commit-id: e54430265f48c151c9848fc45604c45ce6108e02
Former-commit-id: a97f1a69f63dda3373aa0e5088fbc44e9a57dbcd

+++ backend/plugin/db/mysql/mysql.go
@@ -310,7 +310,7 @@ func (d *Driver) Execute(ctx context.Context, statement string, opts db.ExecuteO
 
 	var totalCommands int
 	var commands []base.SingleSQL
-	var originalIndex []int
+	var originalIndex []int32
 	if len(statement) <= common.MaxSheetCheckSize {
 		singleSQLs, err := mysqlparser.SplitSQL(statement)
 		if err != nil {
@@ -328,7 +328,7 @@ func (d *Driver) Execute(ctx context.Context, statement string, opts db.ExecuteO
 				Text: statement,
 			},
 		}
-		originalIndex = []int{0}
+		originalIndex = []int32{0}
 	}
 
 	tx, err := conn.BeginTx(ctx, nil)
@@ -367,7 +367,7 @@ func (d *Driver) Execute(ctx context.Context, statement string, opts db.ExecuteO
 				})
 			}
 
-			indexes := []int32{int32(originalIndex[i])}
+			indexes := []int32{originalIndex[i]}
 			opts.LogCommandExecute(indexes)
 
 			sqlResult, err := exer.ExecContext(ctx, command.Text, nil)
